feat: rename WebTolgee to Tolgee
