+++ packages/web/rollup.config.js
@@ -6,7 +6,7 @@ export default [
   ...buildVanilla('src/index.ts', 'web'),
   ...buildVanilla('src/InContextTools.ts', 'in-context-tools'),
   ...buildVanilla('src/ObserverPlugin.ts', 'observer-plugin'),
-  ...buildVanilla('src/WebTolgee.ts', 'web-tolgee'),
+  ...buildVanilla('src/Tolgee.ts', 'tolgee'),
   ...buildVanilla('src/LanguageStorage.ts', 'language-storage'),
   ...buildVanilla('src/LanguageDetector.ts', 'language-detector'),
   ...buildVanilla('src/BackendFetch.ts', 'backend-fetch'),
